% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sdb_download}
\alias{sdb_download}
\title{Download All Dataset Files to Local Directory by DOI}
\usage{
sdb_download(doi, user_path)
}
\arguments{
\item{doi}{A character string specifying the DOI (e.g., "10.57760/sciencedb.15594").}

\item{user_path}{A character string specifying the local directory path where downloaded files will be saved.}
}
\value{
Messages informing users the number of files saved successfully to the specified user path.
}
\description{
Download All Dataset Files to Local Directory by DOI
}
\examples{
\dontrun{
sdb_download(doi = "10.57760/sciencedb.15594",
             user_path = "~/Desktop")
}
}
